requirements.txt
requests
dotenv
pillow
groq
gradio
